+++ client/src/components/Landing/landing.css
@@ -3,10 +3,10 @@
     max-width: 1600px;
     position: relative;
   }
-
+/* 
   .swiper .swiper-slide {
     width: 250px;
-  }
+  } */
 
   .swiper img {
     height: 255px;
